refactor(cli): isolate analytics worker bootstrap
